src/main/java/cc/tonny/orderservice/adapter/inbound/OrderController.java
package cc.tonny.orderservice.adapter.inbound;

import cc.tonny.orderservice.adapter.inbound.dto.OrderRequest;
import cc.tonny.orderservice.domain.Order;
import cc.tonny.orderservice.domain.OrderService;
import jakarta.validation.Valid;
import lombok.AllArgsConstructor;
import org.springframework.web.bind.annotation.*;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;

@RestController
@RequestMapping("orders")
@AllArgsConstructor
public class OrderController {
    private final OrderService orderService;

    @GetMapping
    public Flux<Order> getAllOrders() {
        return orderService.getAllOrders();
    }

    @PostMapping
    public Mono<Order> submitOrder(@RequestBody @Valid OrderRequest orderRequest) {
        return orderService.submitOrder(orderRequest.isbn(), orderRequest.quantity());
    }
}
